design: nav,profile
